admin member 정상/탈퇴 회원 구분 추가

+++ src/main/java/daangnmungcat/dto/PageMaker.java
@@ -55,7 +55,7 @@ public class PageMaker {
 	}
 
 	private void calcData() {
-		lastPage = (int) (Math.ceil(totalCount / cri.getPerPageNum()));
+		lastPage = (int) (Math.ceil(totalCount / (double)cri.getPerPageNum()));
 		endPage = (int) (Math.ceil(cri.getPage() / (double) displayPageNum) * displayPageNum);
 		startPage = (endPage - displayPageNum) + 1;
